+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/AuxData.hs
@@ -9,10 +9,16 @@ module Cardano.Ledger.Api.Tx.AuxData (
 
   -- * Alonzo
   AlonzoTxAuxData (..),
+  mkAlonzoTxAuxData,
+  getAlonzoTxAuxDataScripts,
 )
 where
 
 import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..))
-import Cardano.Ledger.Alonzo.TxAuxData (AlonzoTxAuxData (..))
+import Cardano.Ledger.Alonzo.TxAuxData (
+  AlonzoTxAuxData (..),
+  getAlonzoTxAuxDataScripts,
+  mkAlonzoTxAuxData,
+ )
 import Cardano.Ledger.Core (EraTxAuxData (..))
 import Cardano.Ledger.Shelley.TxAuxData (ShelleyTxAuxData (..))